+++ user/zone/ml_traveltime_vconstant.c
@@ -42,10 +42,11 @@ typedef struct twod {
 
 
 /*Traveltime functions for constant velocity---------------------------------------------------------------------------*/
-float T0_k(twod y_k,twod y_k1)
+
+double T0_k(twod y_k,twod y_k1)
 /*<Traveltime>*/
 {
-	float t_k;
+	double t_k;
 	
 	t_k = hypotf(y_k.x-y_k1.x,y_k.z-y_k1.z)/y_k.v2;
 	
@@ -53,10 +54,10 @@ float T0_k(twod y_k,twod y_k1)
 	
 }
 
-float T0_k_k(twod y_k, twod y_k1) 
+double T0_k_k(twod y_k, twod y_k1) 
 /*<Derivative of T with respect to x_k>*/
 {
-	float t_k_k;
+	double t_k_k;
 	
 	t_k_k = (y_k.x-y_k1.x)/(y_k.v2*hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x));
 	
@@ -64,10 +65,10 @@ float T0_k_k(twod y_k, twod y_k1)
 	
 }
 
-float T0_k_k1(twod y_k, twod y_k1) 
+double T0_k_k1(twod y_k, twod y_k1) 
 /*<Derivative of T with respect to x_k1>*/
 {
-	float t_k_k1;
+	double t_k_k1;
 	
 	t_k_k1 = (y_k1.x-y_k.x)/(y_k.v2*hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x));
 	
@@ -75,10 +76,10 @@ float T0_k_k1(twod y_k, twod y_k1)
 	
 }
 
-float T0_k_k_k(twod y_k, twod y_k1)
+double T0_k_k_k(twod y_k, twod y_k1)
 /*<Second derivative of T with respect to x_k>*/
 {
-	float t_k_k_k;
+	double t_k_k_k;
 	
 	t_k_k_k = 1/(y_k.v2*hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x)) - (pow(y_k.x-y_k1.x,2))/(y_k.v2*pow(hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x),3));
 	
@@ -86,10 +87,10 @@ float T0_k_k_k(twod y_k, twod y_k1)
 	
 }
 
-float T0_k_k1_k1(twod y_k, twod y_k1)  
+double T0_k_k1_k1(twod y_k, twod y_k1)  
 /*<Second derivative of T with respect to x_k1>*/
 {
-	float t_k_k1_k1;
+	double t_k_k1_k1;
 	
 	t_k_k1_k1 = 1/(y_k.v2*hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x)) - (pow(y_k.x-y_k1.x,2))/(y_k.v2*pow(hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x),3));
 				 
@@ -97,10 +98,10 @@ float T0_k_k1_k1(twod y_k, twod y_k1)
 	
 }
 
-float T0_k_k_k1(twod y_k, twod y_k1) 
+double T0_k_k_k1(twod y_k, twod y_k1) 
 /*<Second derivative of T with respect to x_k and x_k1>*/
 {
-	float t_k_k_k1;
+	double t_k_k_k1;
 	
 	t_k_k_k1 = (-1)*(1/(y_k.v2*hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x)) - (pow(y_k.x-y_k1.x,2))/(y_k.v2*pow(hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x),3)));
 					 
@@ -108,10 +109,10 @@ float T0_k_k_k1(twod y_k, twod y_k1)
 	
 }
 
-float T0_k_zk(twod y_k, twod y_k1)  
+double T0_k_zk(twod y_k, twod y_k1)  
 /*<Derivative of T with respect to z_k>*/
 {
-	float t_k_zk;
+	double t_k_zk;
 	
 	t_k_zk = (y_k.z-y_k1.z)/(y_k.v2*hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x));
 	
@@ -119,10 +120,10 @@ float T0_k_zk(twod y_k, twod y_k1)
 	
 }
 
-float T0_k_zk1(twod y_k, twod y_k1) 
+double T0_k_zk1(twod y_k, twod y_k1) 
 /*<Derivative of T with respect to z_k1>*/
 {
-	float t_k_zk1;
+	double t_k_zk1;
 	
 	t_k_zk1 = (y_k1.z-y_k.z)/(y_k.v2*hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x));
 	
@@ -130,10 +131,10 @@ float T0_k_zk1(twod y_k, twod y_k1)
 	
 }
 
-float T0_k_zk_zk(twod y_k, twod y_k1)  
+double T0_k_zk_zk(twod y_k, twod y_k1)  
 /*<Second Derivative of T with respect to z_k>*/
 {
-	float t_k_zk_zk;
+	double t_k_zk_zk;
 	
 	t_k_zk_zk = 1/(y_k.v2*hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x)) - (pow(y_k.z-y_k1.z,2))/(y_k.v2*pow(hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x),3));
 	
@@ -141,10 +142,10 @@ float T0_k_zk_zk(twod y_k, twod y_k1)
 	
 }
 
-float T0_k_zk1_zk1(twod y_k, twod y_k1)  
+double T0_k_zk1_zk1(twod y_k, twod y_k1)  
 /*<Second Derivative of T with respect to z_k1>*/
 {
-	float t_k_zk1_zk1;
+	double t_k_zk1_zk1;
 	
 	t_k_zk1_zk1 = 1/(y_k.v2*hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x)) - (pow(y_k.z-y_k1.z,2))/(y_k.v2*pow(hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x),3));
 	
@@ -152,10 +153,10 @@ float T0_k_zk1_zk1(twod y_k, twod y_k1)
 	
 }
 
-float T0_k_zk_zk1(twod y_k, twod y_k1)  
+double T0_k_zk_zk1(twod y_k, twod y_k1)  
 /*<Second Derivative of T with respect to z_k and z_k1>*/
 {
-	float t_k_zk_zk1;
+	double t_k_zk_zk1;
 	
 	t_k_zk_zk1 = (pow(y_k.z-y_k1.z,2))/(y_k.v2*pow(hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x),3)) -  1/(y_k.v2*hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x));
 	
@@ -163,10 +164,10 @@ float T0_k_zk_zk1(twod y_k, twod y_k1)
 	
 }
 
-float T0_k_k_zk(twod y_k, twod y_k1)  
+double T0_k_k_zk(twod y_k, twod y_k1)  
 /*<Second derivative of T with respect to x_k and z_k>*/
 {
-	float t_k_k_zk;
+	double t_k_k_zk;
 	
 	t_k_k_zk =  ((y_k1.z-y_k.z)*(y_k.x-y_k1.x))/(y_k.v2*pow(hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x),3));	
 	
@@ -174,10 +175,10 @@ float T0_k_k_zk(twod y_k, twod y_k1)
 	
 }
 
-float T0_k_k1_zk1(twod y_k, twod y_k1) 
+double T0_k_k1_zk1(twod y_k, twod y_k1) 
 /*<Second derivative of T with respect to x_k1 and z_k1>*/
 {
-	float t_k_k1_zk1;
+	double t_k_k1_zk1;
 	
 	t_k_k1_zk1 = ((y_k1.z-y_k.z)*(y_k.x-y_k1.x))/(y_k.v2*pow(hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x),3));	
 	
@@ -185,10 +186,10 @@ float T0_k_k1_zk1(twod y_k, twod y_k1)
 	
 }
 
-float T0_k_k_zk1(twod y_k, twod y_k1)  
+double T0_k_k_zk1(twod y_k, twod y_k1)  
 /*<Second derivative of T with respect to x_k and z_k1>*/
 {
-	float t_k_k_zk1;
+	double t_k_k_zk1;
 	
 	t_k_k_zk1 = (-1)*(((y_k1.z-y_k.z)*(y_k.x-y_k1.x))/(y_k.v2*pow(hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x),3)));	
 	
@@ -196,10 +197,10 @@ float T0_k_k_zk1(twod y_k, twod y_k1)
 	
 }
 
-float T0_k_k1_zk(twod y_k, twod y_k1)  
+double T0_k_k1_zk(twod y_k, twod y_k1)  
 /*<Second derivative of T with respect to x_k1 and z_k>*/
 {
-	float t_k_k1_zk;
+	double t_k_k1_zk;
 	
 	t_k_k1_zk = (-1)*(((y_k1.z-y_k.z)*(y_k.x-y_k1.x))/(y_k.v2*pow(hypotf(y_k.z-y_k1.z,y_k.x-y_k1.x),3)));
